app/requirements.txt
tensorflow-gpu
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api
matplotlib
numpy